tests: split the aio net tests by what each one is about
net.rs reached 1479 lines, three times the reading budget, one justified
increment at a time. Nobody reads it end to end, so nobody edits it from
an understanding of what is already there — which is how it grew.

Four subjects, four files, no test changed:

  net.rs        accept and connect on the default backend
  netcancel.rs  a cancelled pool operation ends, one per operation
  netend.rs     the endings nobody asked for: a close, a deadline, a retire
  netread.rs    a read counts the remainder its port already holds

The socket fixtures two of them share — set_nonblocking,
full_backlog_listener, fill_tcp_backlog — move to the tests module beside
temp_path and assert_cancel_retires. stream_socket_pair has one caller and
goes with it.

The 62 aio tests pass, and the set is unchanged.
