+++ src/io/aio/tests/net.rs
@@ -1,3 +1,8 @@
+// audited: 2026-09-05
+// src/io/AGENTS.md
+//! Accept and connect on the platform's default backend, including both in
+//! flight on one ring at once.
+
 use super::*;
 
 /// Regression test: wait() must not return 0 completions when an accept
@@ -413,1059 +418,5 @@ fn test_accept_and_connect_concurrent() {
         let ids: Vec<SubmissionId> = all.iter().map(|c| c.id).collect();
         assert!(ids.contains(&accept_id), "missing accept");
         assert!(ids.contains(&connect_id), "missing connect");
-    });
-}
-
-/// Take a descriptor non-blocking, whatever the platform spells the flag.
-/// `SOCK_NONBLOCK` is a Linux extension to `socket(2)`; `fcntl` is everywhere.
-fn set_nonblocking(fd: libc::c_int) {
-    unsafe {
-        let flags = libc::fcntl(fd, libc::F_GETFL);
-        assert!(flags >= 0, "F_GETFL failed");
-        assert_eq!(libc::fcntl(fd, libc::F_SETFL, flags | libc::O_NONBLOCK), 0);
-    }
-}
-
-/// A loopback TCP listener with a backlog of one, and the port it took.
-///
-/// The small backlog is the point: paired with `fill_tcp_backlog`, it gives a
-/// listener the kernel refuses further connections to, which is a peer that
-/// never answers without needing a network to reach one.
-fn full_backlog_listener() -> (libc::c_int, u16) {
-    unsafe {
-        let fd = libc::socket(libc::AF_INET, libc::SOCK_STREAM, 0);
-        assert!(fd >= 0, "socket() failed");
-        let mut addr: libc::sockaddr_in = std::mem::zeroed();
-        addr.sin_family = libc::AF_INET as libc::sa_family_t;
-        addr.sin_port = 0;
-        addr.sin_addr.s_addr = u32::from(std::net::Ipv4Addr::LOCALHOST).to_be();
-        assert_eq!(
-            libc::bind(
-                fd,
-                &addr as *const _ as *const libc::sockaddr,
-                std::mem::size_of::<libc::sockaddr_in>() as libc::socklen_t
-            ),
-            0,
-            "bind() failed"
-        );
-        assert_eq!(libc::listen(fd, 1), 0, "listen() failed");
-        let mut bound: libc::sockaddr_in = std::mem::zeroed();
-        let mut len = std::mem::size_of::<libc::sockaddr_in>() as libc::socklen_t;
-        libc::getsockname(fd, &mut bound as *mut _ as *mut libc::sockaddr, &mut len);
-        (fd, u16::from_be(bound.sin_port))
-    }
-}
-
-/// Fill the accept queue of the listener on `port`, and return the descriptors
-/// that hold it full. Nobody accepts these, so the kernel has nowhere to put a
-/// further connection and drops its SYNs.
-///
-/// The caller holds the descriptors for the length of its test: closing one
-/// frees a queue slot, and the connect under test would then complete.
-fn fill_tcp_backlog(port: u16) -> Vec<libc::c_int> {
-    let mut queued = Vec::new();
-    for _ in 0..8 {
-        let c = unsafe { libc::socket(libc::AF_INET, libc::SOCK_STREAM, 0) };
-        assert!(c >= 0, "socket() failed");
-        set_nonblocking(c);
-        unsafe {
-            let mut addr: libc::sockaddr_in = std::mem::zeroed();
-            addr.sin_family = libc::AF_INET as libc::sa_family_t;
-            addr.sin_port = port.to_be();
-            addr.sin_addr.s_addr = u32::from(std::net::Ipv4Addr::LOCALHOST).to_be();
-            libc::connect(
-                c,
-                &addr as *const _ as *const libc::sockaddr,
-                std::mem::size_of::<libc::sockaddr_in>() as libc::socklen_t,
-            );
-        }
-        queued.push(c);
-    }
-    queued
-}
-
-/// A cancelled accept must END on the thread-pool backend, not be abandoned.
-///
-/// Closing a listener under a parked accept is how a program reaches the state
-/// `assert_cancel_retires` describes: an entry whose worker is gone for good.
-///
-/// Built on `new_thread_pool` rather than `AsyncBackend::new` on purpose: on a
-/// Linux host with io_uring the default backend is the ring, and this property
-/// would go unchecked on every dev box while only CI (and every non-Linux
-/// build, which has no other arm) ran the code it is about.
-#[test]
-fn a_cancelled_pool_accept_ends_rather_than_being_abandoned() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-        use std::os::unix::io::FromRawFd;
-
-        // A BLOCKING listener, deliberately. With SOCK_NONBLOCK the worker's
-        // `accept` returns EAGAIN at once and the operation ends whatever the
-        // cancel path does — the test would pass without ever putting a worker
-        // in the blocking `accept` this is about.
-        let listener_fd = unsafe {
-            let fd = libc::socket(libc::AF_INET, libc::SOCK_STREAM, 0);
-            assert!(fd >= 0, "socket() failed");
-            let mut addr: libc::sockaddr_in = std::mem::zeroed();
-            addr.sin_family = libc::AF_INET as libc::sa_family_t;
-            addr.sin_port = 0;
-            addr.sin_addr.s_addr = u32::from(std::net::Ipv4Addr::LOCALHOST).to_be();
-            assert_eq!(
-                libc::bind(
-                    fd,
-                    &addr as *const _ as *const libc::sockaddr,
-                    std::mem::size_of::<libc::sockaddr_in>() as libc::socklen_t
-                ),
-                0
-            );
-            assert_eq!(libc::listen(fd, 128), 0);
-            fd
-        };
-        let listener_port = h.ctx().external(
-            "port",
-            Port::new_tcp_listener(
-                unsafe { std::os::unix::io::OwnedFd::from_raw_fd(listener_fd) },
-                "127.0.0.1:0".to_string(),
-            ),
-        );
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let accept_port_val = h.ctx().external(
-            "port",
-            Port::new_unopened(
-                PortKind::TcpStream,
-                Direction::ReadWrite,
-                Encoding::Binary,
-                String::new(),
-            ),
-        );
-        let accept_id = backend
-            .submit(
-                &IoRequest {
-                    op: PortOp::Accept {
-                        options: Default::default(),
-                        encoding: crate::port::Encoding::Binary,
-                        accept_port: accept_port_val,
-                    }
-                    .into(),
-                    port: listener_port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        assert_cancel_retires(&backend, accept_id, "accept");
-    });
-}
-
-/// Closing a listener must END its parked pool accept, on every platform.
-///
-/// `port/close` is the only unblocking mechanism a program has for an accept
-/// nobody cancels — an accept loop parked in a live process, closed by another
-/// process at teardown (tests/elle/process-accept-close.lisp is the scheduler
-/// shape). The close path may not lean on `shutdown(2)` for this: shutdown of
-/// a LISTENING socket is a Linux extension — macOS and the BSDs return
-/// ENOTCONN and wake nothing, and the accept's worker then polls the retired
-/// descriptor forever while the scheduler waits on a completion that never
-/// comes. The wake must come from the operation's stop pipe instead.
-///
-/// Built on `new_thread_pool` for the reason the cancellation tests give: on a
-/// Linux dev box the default backend is the ring, and this property would go
-/// unchecked everywhere it can regress.
-#[test]
-fn closing_a_listener_ends_its_parked_pool_accept() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-        use std::os::unix::io::FromRawFd;
-
-        // A BLOCKING listener, deliberately — see the cancellation test above.
-        let listener_fd = unsafe {
-            let fd = libc::socket(libc::AF_INET, libc::SOCK_STREAM, 0);
-            assert!(fd >= 0, "socket() failed");
-            let mut addr: libc::sockaddr_in = std::mem::zeroed();
-            addr.sin_family = libc::AF_INET as libc::sa_family_t;
-            addr.sin_port = 0;
-            addr.sin_addr.s_addr = u32::from(std::net::Ipv4Addr::LOCALHOST).to_be();
-            assert_eq!(
-                libc::bind(
-                    fd,
-                    &addr as *const _ as *const libc::sockaddr,
-                    std::mem::size_of::<libc::sockaddr_in>() as libc::socklen_t
-                ),
-                0
-            );
-            assert_eq!(libc::listen(fd, 128), 0);
-            fd
-        };
-        let listener_port = h.ctx().external(
-            "port",
-            Port::new_tcp_listener(
-                unsafe { std::os::unix::io::OwnedFd::from_raw_fd(listener_fd) },
-                "127.0.0.1:0".to_string(),
-            ),
-        );
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let accept_port_val = h.ctx().external(
-            "port",
-            Port::new_unopened(
-                PortKind::TcpStream,
-                Direction::ReadWrite,
-                Encoding::Binary,
-                String::new(),
-            ),
-        );
-        let accept_id = backend
-            .submit(
-                &IoRequest {
-                    op: PortOp::Accept {
-                        options: Default::default(),
-                        encoding: crate::port::Encoding::Binary,
-                        accept_port: accept_port_val,
-                    }
-                    .into(),
-                    port: listener_port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        // Let the worker reach its wait before the close arrives.
-        let mut submitted = false;
-        for _ in 0..200 {
-            if backend.workers() > 0 {
-                submitted = true;
-                break;
-            }
-            std::thread::sleep(std::time::Duration::from_millis(5));
-        }
-        assert!(submitted, "the pool never took the accept out to a worker");
-        std::thread::sleep(std::time::Duration::from_millis(50));
-
-        // Close the listener. The close itself completes immediately; the
-        // parked accept must then complete too — as an error, within a bound.
-        let close_id = backend
-            .submit(
-                &IoRequest {
-                    op: IoOp::Close,
-                    port: listener_port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        let mut accept_completion = None;
-        for _ in 0..40 {
-            for c in backend.wait(50).unwrap() {
-                if c.id == accept_id {
-                    accept_completion = Some(c);
-                } else {
-                    assert_eq!(c.id, close_id, "unexpected completion");
-                }
-            }
-            if accept_completion.is_some() {
-                break;
-            }
-        }
-        let accept_completion = accept_completion.expect(
-            "the parked accept never completed after its listener closed — \
-             the fiber waiting on it would wait forever",
-        );
-        assert!(
-            accept_completion.result.is_err(),
-            "an accept on a closed listener must not report success",
-        );
-        assert_eq!(
-            backend.workers(),
-            0,
-            "the accept's worker never came back after the close",
-        );
-        assert!(
-            !backend.has_pending(),
-            "the accept is still pending after the close",
-        );
-    });
-}
-
-/// A cancelled datagram receive must END on the thread-pool backend.
-///
-/// The accept test's twin on the other open-ended socket operation: a socket
-/// nobody sends to waits exactly as long as a listener nobody calls. `ev/timeout`
-/// around a `udp/recv-from` is the caller that meets it — `lib/dns.lisp` sends a
-/// query and waits for a reply that a lossy network need never deliver.
-///
-/// The socket is deliberately BLOCKING, for the reason the accept test gives:
-/// a non-blocking one returns EAGAIN at once and the operation ends whatever
-/// the cancel path does.
-#[test]
-fn a_cancelled_pool_recvfrom_ends_rather_than_being_abandoned() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-        use std::os::unix::io::FromRawFd;
-
-        let sock_fd = unsafe {
-            let fd = libc::socket(libc::AF_INET, libc::SOCK_DGRAM, 0);
-            assert!(fd >= 0, "socket() failed");
-            let mut addr: libc::sockaddr_in = std::mem::zeroed();
-            addr.sin_family = libc::AF_INET as libc::sa_family_t;
-            addr.sin_port = 0;
-            addr.sin_addr.s_addr = u32::from(std::net::Ipv4Addr::LOCALHOST).to_be();
-            assert_eq!(
-                libc::bind(
-                    fd,
-                    &addr as *const _ as *const libc::sockaddr,
-                    std::mem::size_of::<libc::sockaddr_in>() as libc::socklen_t
-                ),
-                0
-            );
-            fd
-        };
-        let sock_port = h.ctx().external(
-            "port",
-            Port::new_udp_socket(
-                unsafe { std::os::unix::io::OwnedFd::from_raw_fd(sock_fd) },
-                "127.0.0.1:0".to_string(),
-            ),
-        );
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let recv_id = backend
-            .submit(
-                &IoRequest {
-                    // The pool worker receives into its own buffer, so the
-                    // destination struct a fiber would pass is not needed here.
-                    op: PortOp::RecvFrom {
-                        count: 64,
-                        result: Value::NIL,
-                    }
-                    .into(),
-                    port: sock_port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        assert_cancel_retires(&backend, recv_id, "recvfrom");
-    });
-}
-
-/// A cancelled write must END on the thread-pool backend.
-///
-/// The full-write invariant makes a write run to the end of its payload, and a
-/// payload larger than the send buffer only gets there as the peer takes what is
-/// already in it. So a write waits on a peer exactly as an accept and a
-/// `recvfrom` do, and a peer that stops reading is the wait that never ends.
-/// `ev/timeout` around a `port/write` is the caller that meets it.
-///
-/// The trap: the peer socket must stay OPEN. Closing it ends the write with
-/// `EPIPE` for a reason that has nothing to do with cancellation, and the test
-/// would then pass with the stop pipe taken away again.
-///
-/// Counter-factual: with the write submitted under `Bounds::new(timeout, None)`,
-/// `OpBound::new` finds neither a deadline nor a stop to enforce and leaves the
-/// descriptor blocking, so the worker sits inside `write(2)` where no poll and no
-/// stop can reach it. `assert_cancel_retires` then fails on the worker that never
-/// comes back.
-#[test]
-fn a_cancelled_pool_write_ends_rather_than_being_abandoned() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-        use std::os::unix::io::FromRawFd;
-
-        // A socketpair is the smallest peer that can stop reading: one end is
-        // the port, the other is a peer this test simply never touches.
-        let mut fds = [0 as libc::c_int; 2];
-        assert_eq!(
-            unsafe { libc::socketpair(libc::AF_UNIX, libc::SOCK_STREAM, 0, fds.as_mut_ptr()) },
-            0,
-            "socketpair() failed"
-        );
-        let (write_fd, peer_fd) = (fds[0], fds[1]);
-
-        // Shrink the send buffer so the payload below cannot fit in it whatever
-        // the platform's default happens to be. The kernel is free to round this
-        // up, which is why the payload is larger than any plausible rounding.
-        let want: libc::c_int = 4096;
-        unsafe {
-            libc::setsockopt(
-                write_fd,
-                libc::SOL_SOCKET,
-                libc::SO_SNDBUF,
-                &want as *const _ as *const libc::c_void,
-                std::mem::size_of::<libc::c_int>() as libc::socklen_t,
-            );
-            libc::setsockopt(
-                peer_fd,
-                libc::SOL_SOCKET,
-                libc::SO_RCVBUF,
-                &want as *const _ as *const libc::c_void,
-                std::mem::size_of::<libc::c_int>() as libc::socklen_t,
-            );
-        }
-
-        let port = h.ctx().external(
-            "port",
-            Port::new_unix_stream(
-                unsafe { std::os::unix::io::OwnedFd::from_raw_fd(write_fd) },
-                "socketpair".to_string(),
-            ),
-        );
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let data = h.ctx().bytes(vec![b'x'; 4 << 20]);
-        let write_id = backend
-            .submit(
-                &IoRequest {
-                    // No `:timeout`: the deadline is the ending this test must
-                    // not be able to reach for. Cancellation is the only one
-                    // left, and it is the one under test.
-                    op: PortOp::Write { data }.into(),
-                    port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        assert_cancel_retires(&backend, write_id, "write");
-
-        unsafe { libc::close(peer_fd) };
-    });
-}
-
-/// A cancelled TCP connect must END on the thread-pool backend.
-///
-/// The stall is a listener whose accept queue is full: the kernel drops further
-/// SYNs, so the handshake never completes and the connect waits on a peer that
-/// will not answer. A blocking `connect(2)` holds its worker through the whole
-/// SYN-retry sequence — minutes, with no way for a cancel to reach it.
-#[test]
-fn a_cancelled_pool_tcp_connect_ends_rather_than_being_abandoned() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-
-        let (listener_fd, bound_port) = full_backlog_listener();
-        let queued = fill_tcp_backlog(bound_port);
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let connect_port = h.ctx().external(
-            "port",
-            Port::new_unopened(
-                PortKind::TcpStream,
-                Direction::ReadWrite,
-                Encoding::Binary,
-                format!("127.0.0.1:{}", bound_port),
-            ),
-        );
-        let connect_id = backend
-            .submit(
-                &IoRequest {
-                    op: IoOp::Connect {
-                        addr: crate::io::request::ConnectAddr::Tcp {
-                            addr: "127.0.0.1".parse().unwrap(),
-                            port: bound_port,
-                            options: Default::default(),
-                            encoding: crate::port::Encoding::Binary,
-                        },
-                    },
-                    port: connect_port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        assert_cancel_retires(&backend, connect_id, "tcp connect");
-
-        for c in queued {
-            unsafe { libc::close(c) };
-        }
-        unsafe { libc::close(listener_fd) };
-    });
-}
-
-/// A pool connect must stop at the caller's `:timeout`, and say so.
-///
-/// The same full accept queue as the cancellation test above, waited on with a
-/// deadline instead of cancelled. Two things are pinned: the connect ends near
-/// its deadline rather than at the kernel's own, minutes later; and it reports
-/// `:timeout`, the kind `ev/timeout` and every caller that distinguishes a
-/// deadline from a broken connection matches on.
-#[test]
-fn a_pool_connect_reports_its_own_deadline_as_a_timeout() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-
-        let (listener_fd, bound_port) = full_backlog_listener();
-        let queued = fill_tcp_backlog(bound_port);
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let connect_port = h.ctx().external(
-            "port",
-            Port::new_unopened(
-                PortKind::TcpStream,
-                Direction::ReadWrite,
-                Encoding::Binary,
-                format!("127.0.0.1:{}", bound_port),
-            ),
-        );
-        let started = std::time::Instant::now();
-        let connect_id = backend
-            .submit(
-                &IoRequest {
-                    op: IoOp::Connect {
-                        addr: crate::io::request::ConnectAddr::Tcp {
-                            addr: "127.0.0.1".parse().unwrap(),
-                            port: bound_port,
-                            options: Default::default(),
-                            encoding: crate::port::Encoding::Binary,
-                        },
-                    },
-                    port: connect_port,
-                    timeout: Some(std::time::Duration::from_millis(200)),
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        let mut completions = Vec::new();
-        for _ in 0..40 {
-            completions.extend(backend.wait(200).unwrap());
-            if !completions.is_empty() {
-                break;
-            }
-        }
-        assert_eq!(
-            completions.len(),
-            1,
-            "the connect never completed within 8s of a 200ms deadline",
-        );
-        assert_eq!(completions[0].id, connect_id);
-        let elapsed = started.elapsed();
-        assert!(
-            elapsed < std::time::Duration::from_secs(2),
-            "the connect took {:?} against a 200ms deadline — it waited on the \
-             kernel's own retry sequence instead of its own bound",
-            elapsed,
-        );
-
-        let err = completions[0]
-            .result
-            .as_ref()
-            .expect_err("a connect to a full accept queue must not succeed");
-        let fields = err.as_struct().expect("an io error is a struct");
-        assert_eq!(
-            crate::value::sorted_struct_get(fields, &TableKey::keyword("error")).copied(),
-            Some(crate::value::Value::keyword("timeout")),
-            "a connect that ran out its deadline must report :timeout, not a \
-             generic :io-error — `ev/timeout` and `timed-out?` match on the kind",
-        );
-
-        for c in queued {
-            unsafe { libc::close(c) };
-        }
-        unsafe { libc::close(listener_fd) };
-    });
-}
-
-/// A cancelled Unix connect must END on the thread-pool backend.
-///
-/// AF_UNIX reports a full backlog differently from TCP, which is why it is
-/// pinned separately: a non-blocking connect returns EAGAIN with no readiness
-/// to poll for, so the operation paces its retries and watches for the stop
-/// between them. A blocking one waits inside the kernel until the listener
-/// accepts, which a listener that has stopped accepting never does.
-///
-/// Linux-only: macOS and the BSDs report a full backlog as `ECONNREFUSED`,
-/// which the connect does not pace, so a parked Unix connect cannot be
-/// arranged there.
-#[test]
-#[cfg(target_os = "linux")]
-fn a_cancelled_pool_unix_connect_ends_rather_than_being_abandoned() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-
-        let path = temp_path("unix-connect-cancel");
-        let (sun, addr_len) = crate::io::sockaddr::build_unix(&path).unwrap();
-        let listener_fd = unsafe {
-            let fd = libc::socket(libc::AF_UNIX, libc::SOCK_STREAM, 0);
-            assert!(fd >= 0, "socket() failed");
-            assert_eq!(
-                libc::bind(fd, &sun as *const _ as *const libc::sockaddr, addr_len),
-                0,
-                "bind({}) failed",
-                path
-            );
-            assert_eq!(libc::listen(fd, 1), 0);
-            fd
-        };
-
-        // Fill the backlog. AF_UNIX says so directly — a non-blocking connect
-        // to a full queue reports EAGAIN — so the setup can prove the connect
-        // under test really has nothing to complete against.
-        let mut queued: Vec<libc::c_int> = Vec::new();
-        let mut full = false;
-        for _ in 0..8 {
-            let c = unsafe { libc::socket(libc::AF_UNIX, libc::SOCK_STREAM, 0) };
-            assert!(c >= 0, "socket() failed");
-            set_nonblocking(c);
-            let r =
-                unsafe { libc::connect(c, &sun as *const _ as *const libc::sockaddr, addr_len) };
-            if r == 0 {
-                queued.push(c);
-                continue;
-            }
-            let errno = std::io::Error::last_os_error().raw_os_error().unwrap_or(1);
-            unsafe { libc::close(c) };
-            assert!(
-                errno == libc::EAGAIN || errno == libc::EWOULDBLOCK,
-                "connect({}) failed with errno {}, expected EAGAIN/EWOULDBLOCK",
-                path,
-                errno
-            );
-            full = true;
-            break;
-        }
-        assert!(
-            full,
-            "the listener's backlog never filled, so the connect under test \
-             would complete instead of waiting"
-        );
-
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-        let connect_port = h.ctx().external(
-            "port",
-            Port::new_unopened(
-                PortKind::UnixStream,
-                Direction::ReadWrite,
-                Encoding::Binary,
-                path.clone(),
-            ),
-        );
-        let connect_id = backend
-            .submit(
-                &IoRequest {
-                    op: IoOp::Connect {
-                        addr: crate::io::request::ConnectAddr::Unix {
-                            path: path.clone(),
-                            options: Default::default(),
-                            encoding: crate::port::Encoding::Binary,
-                        },
-                    },
-                    port: connect_port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        assert_cancel_retires(&backend, connect_id, "unix connect");
-
-        for c in queued {
-            unsafe { libc::close(c) };
-        }
-        unsafe { libc::close(listener_fd) };
-        let _ = std::fs::remove_file(&path);
-    });
-}
-
-/// A retired accept gives back the connection it took.
-///
-/// An accept that succeeded owns a descriptor: the connection the kernel
-/// handed it. Every other retiring arm gives its descriptor back — a connect
-/// closes the socket it pre-created, an open closes the file it opened — and
-/// an accept must too, or a server whose accept loop ends leaks one socket per
-/// round it had in flight.
-///
-/// The trap: an fd count cannot say this. Tests share a process and run in
-/// parallel, so the number moves under the measurement. The peer can say it
-/// instead — a connection nobody closed leaves the peer's read waiting, while
-/// a closed one ends it.
-#[test]
-fn a_retired_accept_closes_the_connection_it_took() {
-    use std::io::Read;
-    crate::value::arena::with_test_region(|| {
-        for (backend, which) in [
-            (AsyncBackend::new().unwrap(), "the platform default"),
-            (AsyncBackend::new_thread_pool().unwrap(), "the thread pool"),
-        ] {
-            let listener = std::net::TcpListener::bind("127.0.0.1:0").expect("bind");
-            let addr = listener.local_addr().expect("local_addr");
-
-            let heap_ptr = crate::value::arena::leaked_test_heap();
-            // SAFETY: the heap is leaked for the process.
-            let heap = unsafe { &mut *heap_ptr };
-
-            // The listener outlives the operation; only the fiber's own region
-            // goes, which is what makes the entry retire unread.
-            let kept = heap.new_runtime_region();
-            let listener_port = crate::value::build::external(
-                heap,
-                "port",
-                Port::new_tcp_listener(listener.into(), addr.to_string()),
-                kept,
-            );
-            let region = heap.new_runtime_region();
-            let accept_port = crate::value::build::external(
-                heap,
-                "port",
-                Port::new_unopened(
-                    PortKind::TcpStream,
-                    Direction::ReadWrite,
-                    Encoding::Binary,
-                    String::new(),
-                ),
-                region,
-            );
-
-            // The peer connects BEFORE the accept is submitted, so the kernel
-            // has a connection queued and the operation takes one the moment it
-            // runs. An accept retired before it has a connection owns no
-            // descriptor and would close nothing, which is not the case under
-            // test.
-            let client = std::net::TcpStream::connect(addr).expect("connect");
-
-            backend
-                .submit(
-                    &IoRequest {
-                        op: PortOp::Accept {
-                            options: Default::default(),
-                            encoding: Encoding::Binary,
-                            accept_port,
-                        }
-                        .into(),
-                        port: listener_port,
-                        timeout: None,
-                    },
-                    crate::io::pending::Submitter::detached(heap_ptr),
-                )
-                .unwrap();
-
-            // Let the operation reach its completion before the region goes.
-            // Nothing consumes a completion outside `wait`/`poll`, so the answer
-            // sits in the ring or the hub across this settle and is taken below
-            // — after the release, which is the order that makes the entry
-            // retire with a descriptor in hand.
-            wait_for_worker(&backend);
-
-            // The fiber ends: its region goes, and with it the port the accept
-            // would have filled.
-            heap.decref_region(region);
-
-            for _ in 0..40 {
-                let _ = backend.wait(50).unwrap();
-                if !backend.has_pending() && backend.workers() == 0 {
-                    break;
-                }
-            }
-
-            client
-                .set_read_timeout(Some(Duration::from_secs(5)))
-                .expect("set_read_timeout");
-            // The trap: this read is interrupted, not just bounded. The runtime
-            // signals its own threads, and `std::net`'s `read` reports `EINTR`
-            // rather than retrying — so a single call reports "the peer's read
-            // never ended" whatever the descriptor did. Retry until the
-            // descriptor answers or the deadline passes, and let only the
-            // deadline mean the connection is still open.
-            let mut buf = [0u8; 1];
-            let deadline = std::time::Instant::now() + Duration::from_secs(5);
-            let ended = loop {
-                if std::time::Instant::now() >= deadline {
-                    break false;
-                }
-                match (&client).read(&mut buf) {
-                    Ok(0) => break true,
-                    Ok(n) => panic!("{which}: the peer read {n} bytes from a retired accept"),
-                    Err(e) if e.kind() == std::io::ErrorKind::ConnectionReset => break true,
-                    Err(e) if e.kind() == std::io::ErrorKind::Interrupted => continue,
-                    Err(e) if e.kind() == std::io::ErrorKind::WouldBlock => break false,
-                    Err(e) => panic!("{which}: the peer's read failed with {e}"),
-                }
-            };
-            assert!(
-                ended,
-                "{which}: the peer's read never ended — the accept was retired \
-                 but the connection it took was never closed",
-            );
-        }
-    });
-}
-
-/// A connected stream pair. The returned descriptors are the test's to close;
-/// `Port` takes the first, the test writes the reply into the second.
-fn stream_socket_pair() -> (libc::c_int, libc::c_int) {
-    let mut fds = [0 as libc::c_int; 2];
-    assert_eq!(
-        unsafe { libc::socketpair(libc::AF_UNIX, libc::SOCK_STREAM, 0, fds.as_mut_ptr()) },
-        0,
-        "socketpair(2) failed"
-    );
-    (fds[0], fds[1])
-}
-
-/// A `ReadExact` must count the remainder the port is already holding.
-///
-/// A length-prefixed reply arrives as one burst: the `ReadLine` that takes the
-/// header reads a whole chunk, so the body's first bytes come back with it and
-/// the completion stashes them as the port's remainder (`submit.rs` § "A
-/// previous read on this port took more from the kernel than it answered
-/// with"). The `ReadExact` that follows asks for the whole body, and the wire
-/// is short of that count by exactly what the port is holding.
-///
-/// The ring answers this in `uring/drain.rs`: its "enough yet?" test is
-/// `state.buffer.len() + filled + got < count`, so the held bytes count toward
-/// the total and it stops when the wire has delivered the rest. The pool's
-/// runner counts only what it read from the descriptor, so it waits for the
-/// full count from a peer that has already said everything it has to say — and
-/// a redis `GET` of a value past one chunk hangs forever
-/// (`tests/elle/redis-short-read.lisp`, which is gated on a live Redis and so
-/// never runs on the macOS CI box, the only one that uses this backend).
-///
-/// Built on `new_thread_pool` for the reason the rest of this family gives: the
-/// ring is the default on a Linux host, and this property would otherwise go
-/// unchecked everywhere it can regress.
-#[test]
-fn a_pool_read_exact_counts_the_remainder_the_port_already_holds() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-        use std::os::unix::io::FromRawFd;
-
-        let (ours, peer) = stream_socket_pair();
-
-        // Header, body, terminator — one burst, so the header's read
-        // over-reaches into the body. The body is past one 4096-byte
-        // `read_until` chunk, so the `ReadExact` still has to reach the wire.
-        let body_len = 40_000usize;
-        let mut payload = format!("${}\r\n", body_len).into_bytes();
-        payload.extend(std::iter::repeat_n(b'x', body_len));
-        payload.extend_from_slice(b"\r\n");
-
-        // From a thread: a socketpair's send buffer is smaller than this
-        // payload on some platforms (macOS defaults well under it), and a
-        // writer that blocks would deadlock the test rather than fail it.
-        let writer = std::thread::spawn(move || {
-            let mut sent = 0usize;
-            while sent < payload.len() {
-                let n = unsafe {
-                    libc::write(
-                        peer,
-                        payload[sent..].as_ptr() as *const libc::c_void,
-                        payload.len() - sent,
-                    )
-                };
-                assert!(n > 0, "the peer's write failed");
-                sent += n as usize;
-            }
-        });
-
-        let port = h.ctx().external(
-            "port",
-            Port::new_unix_stream(
-                unsafe { std::os::unix::io::OwnedFd::from_raw_fd(ours) },
-                "socketpair".to_string(),
-            ),
-        );
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-
-        let line_id = backend
-            .submit(
-                &IoRequest {
-                    op: PortOp::ReadLine {
-                        buffer: h.ctx().bytes(vec![0u8; 65536]),
-                    }
-                    .into(),
-                    port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-        let mut line_done = false;
-        for _ in 0..40 {
-            for c in backend.wait(50).unwrap() {
-                if c.id == line_id {
-                    line_done = true;
-                }
-            }
-            if line_done {
-                break;
-            }
-        }
-        assert!(line_done, "the header read never completed");
-
-        let count = body_len + 2;
-        let exact_id = backend
-            .submit(
-                &IoRequest {
-                    op: PortOp::ReadExact {
-                        count,
-                        buffer: h.ctx().bytes(vec![0u8; count]),
-                    }
-                    .into(),
-                    port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        // Bounded, because the property under test is that this terminates.
-        let mut exact = None;
-        for _ in 0..80 {
-            for c in backend.wait(50).unwrap() {
-                if c.id == exact_id {
-                    exact = Some(c);
-                }
-            }
-            if exact.is_some() {
-                break;
-            }
-        }
-        let exact = exact.expect(
-            "the body read never completed — it waited for the whole count from \
-             a peer that had already sent everything, because the bytes the \
-             header's read left in the port went uncounted",
-        );
-        assert!(
-            exact.result.is_ok(),
-            "the body read must answer with the body, not an error",
-        );
-
-        writer.join().unwrap();
-        unsafe {
-            libc::close(peer);
-        }
-    });
-}
-
-/// The same, on a TEXT port, where the count is in grapheme clusters.
-///
-/// The binary twin above pins that the remainder is counted at all. This one
-/// pins that it is counted in the port's own unit: a text `ReadExact` asks for
-/// clusters, the remainder holds some of them, and multibyte clusters make the
-/// byte count and the cluster count different numbers. A runner that counted
-/// only what it read — or counted the remainder in bytes — would ask the peer
-/// for clusters it has already sent.
-///
-/// The ring decides this over `state.buffer` joined to the fiber's bytes
-/// (`uring/drain.rs`, the `text_exact` arm), for the reason this runner takes
-/// the remainder as bytes rather than as a length: a cluster can straddle the
-/// boundary between the two, so neither side can be counted alone.
-#[test]
-fn a_pool_text_read_exact_counts_the_remainder_in_clusters() {
-    crate::value::arena::with_test_region(|| {
-        let h = crate::primitives::ctx::TestHeap::new();
-        use std::os::unix::io::FromRawFd;
-
-        let (ours, peer) = stream_socket_pair();
-
-        // Two bytes per cluster, so a byte count and a cluster count can never
-        // be mistaken for each other.
-        let clusters = 20_000usize;
-        let mut payload = b"hdr\n".to_vec();
-        for _ in 0..clusters {
-            payload.extend_from_slice("é".as_bytes());
-        }
-        let writer = std::thread::spawn(move || {
-            let mut sent = 0usize;
-            while sent < payload.len() {
-                let n = unsafe {
-                    libc::write(
-                        peer,
-                        payload[sent..].as_ptr() as *const libc::c_void,
-                        payload.len() - sent,
-                    )
-                };
-                assert!(n > 0, "the peer's write failed");
-                sent += n as usize;
-            }
-        });
-
-        // A pipe port, because the stream constructors fix the encoding at
-        // Binary and the encoding is the whole point here. What the runner
-        // reads is a descriptor either way; the counting unit is what differs.
-        let port = h.ctx().external(
-            "port",
-            Port::new_pipe(
-                unsafe { std::os::unix::io::OwnedFd::from_raw_fd(ours) },
-                Direction::ReadWrite,
-                Encoding::Text,
-                "socketpair".to_string(),
-            ),
-        );
-        let backend = AsyncBackend::new_thread_pool().unwrap();
-
-        let line_id = backend
-            .submit(
-                &IoRequest {
-                    op: PortOp::ReadLine {
-                        buffer: h.ctx().bytes(vec![0u8; 65536]),
-                    }
-                    .into(),
-                    port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-        let mut line_done = false;
-        for _ in 0..40 {
-            for c in backend.wait(50).unwrap() {
-                if c.id == line_id {
-                    line_done = true;
-                }
-            }
-            if line_done {
-                break;
-            }
-        }
-        assert!(line_done, "the header read never completed");
-
-        // A text read reserves four bytes per cluster, as `port/read-exact`
-        // does — the clusters here are two, so the reservation is ample.
-        let exact_id = backend
-            .submit(
-                &IoRequest {
-                    op: PortOp::ReadExact {
-                        count: clusters,
-                        buffer: h.ctx().bytes(vec![0u8; clusters * 4]),
-                    }
-                    .into(),
-                    port,
-                    timeout: None,
-                },
-                crate::io::pending::Submitter::for_test(),
-            )
-            .unwrap();
-
-        let mut exact = None;
-        for _ in 0..80 {
-            for c in backend.wait(50).unwrap() {
-                if c.id == exact_id {
-                    exact = Some(c);
-                }
-            }
-            if exact.is_some() {
-                break;
-            }
-        }
-        let exact = exact.expect(
-            "the cluster read never completed — the clusters the header's read \
-             left in the port went uncounted",
-        );
-        assert!(
-            exact.result.is_ok(),
-            "the cluster read must answer with the text, not an error",
-        );
-
-        writer.join().unwrap();
-        unsafe {
-            libc::close(peer);
-        }
     });
 }
